new subroutine for command parsing, xysets work when no point in win
This code does two runs through the data before plotting to the screen.
First run to get max/min xy values for the data set.  Then 2nd run with
clipping to any xset/yset boundaries and the display turned off.  The
clipping code initializes and updates bounding box data in the PLOTDAT
structure.  The resulting bounding box is sent to the loose_label()
routine to finalize the plot interval.  Finally a third run sends Xlib
drawing commands to the window.   This 2nd run is required to have the window
properly scale itself when NO point is in the viewport, just clipped segments
of lines.